refactor: 이동 정보 관련 기능 클래스 분리
